requirements.txt
discord.py==1.7.1
requests~=2.25.1
discord~=1.0.1
python-decouple==3.4